requirements.txt
pandas>=1.5.0
torch>=1.13.0
torchvision>=0.14.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
flask >=2.0.0
flask-cors>=3.0.10
requests>=2.25.0
gradio>=3.0.0
